rename file to cancel_takeover_unit_tests
